% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_functions.R
\name{adjustcolnames}
\alias{adjustcolnames}
\title{Used internally for formatting the dataframe
Issue if the column names are not 'x' & 'y'}
\usage{
adjustcolnames(d)
}
\description{
Used internally for formatting the dataframe
Issue if the column names are not 'x' & 'y'
}
