% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_functions.R
\name{calcDL}
\alias{calcDL}
\title{Summarises detection limits}
\source{
{
i) J.C. Miller and J.N. Miller (1993), "Statistics for Analytical Chemistry",
3rd ed., Prentice-Hall.
ii) J. Vogelgesang and J. Hädrich (1998), Accred. Qual. Assur., 3:242-255.
iii) A. Hubaux and G. Vos (1970), Anal. Chem., 42:849-855
& D.T. O'Neill, E.A. Rochette and P.J. Ramsay, (2002), Anal. Chem., 74:5907-5911
iv) J. Ranke, (2018), chemCal: Calibration Functions for Analytical Chemistry,
https://CRAN.R-project.org/package=chemCal
}
}
\usage{
calcDL(d, model_type = NULL, dp = NULL)
}
\arguments{
\item{d}{A tibble containing x (concentration) and y (response).}

\item{model_type}{for regression: (l)inear, (q)uadratic or (p)ower.}

\item{dp}{Number of decimal points}
}
\description{
A generic function summarising the detection limits (DLs) according
to linear, quadratic or power regression. For linear regression, the DLs are
estimated using four approaches;i) Miller and Miller, ii) Vogelgesang and Hädrich,
iii) Hubert & Vos, and iv) the R \emph{chemCal} package. The first two DLs are
directly estimated using internal functions, \emph{dl_miller} and \emph{dl_vogelhad}
while the third DL is estimated iteratively using \emph{dl_hubertvos}.
The default regression is assumed to be linear. However, the DLs can also be
estimated for either a quadratic ("q") or power ("p") regression. The user is advised
though to verify the best regression approach by assessment of the residual
plots - see \emph{resid_plot} for further details.

The default number of single decimal point is 1 but this can be changed by the user.
}
\examples{
data(mtbe)
calcDL(mtbe) 	  # single decimal point, with default to linear regression
calcDL(mtbe, 3)  # error as regression type not defined
calcDL(mtbe, dp = 3)

data(chloromethane)
calcDL(chloromethane, dp = 2)
calcDL(chloromethane, "q", dp = 2) # quadratic regression
}
