% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_functions.R
\name{dl_hubertvos}
\alias{dl_hubertvos}
\title{Calculates the detection limit according to Hubert & Vos using iterative calculation}
\usage{
dl_hubertvos(x, y, alpha = NULL, beta = NULL)
}
\arguments{
\item{x}{A vector}

\item{y}{A vector}
}
\description{
Calculates the detection limit according to Hubert & Vos using iterative calculation
}
\examples{
dl_hubertvos(x,y)
}
