% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_functions.R
\name{dl_miller}
\alias{dl_miller}
\title{Calculates the detection limit according to Miller & Miller}
\usage{
dl_miller(x, y)
}
\arguments{
\item{x}{A vector}

\item{y}{A vector}
}
\value{
MMdl
}
\description{
Calculates the detection limit according to Miller & Miller
}
\examples{
dl_miller(x,y)
}
