% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_functions.R
\name{dl_vogelhad}
\alias{dl_vogelhad}
\title{Calculates the detection limit according to Vogelgesang & Hadrich}
\usage{
dl_vogelhad(x, y)
}
\arguments{
\item{x}{A vector}

\item{y}{A vector}
}
\description{
Calculates the detection limit according to Vogelgesang & Hadrich
}
\examples{
dl_vogelhad(x,y)
}
