% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_functions.R
\name{least_sq_reg}
\alias{least_sq_reg}
\title{Used internally for calculating the least square regression coefficients
of model - assumes a linear response}
\usage{
least_sq_reg(x, y)
}
\description{
Used internally for calculating the least square regression coefficients
of model - assumes a linear response
}
