% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mtbe}
\alias{mtbe}
\title{Calibration dataset for methyl tert-butyl ether in water using GC-MS.}
\format{
A data frame with 18 rows and 2 variables:
\describe{
  \item{concentration}{μg/L}
  \item{area}{arbitrary units}
}
}
\source{
{D.T. O'Neill, E.A. Rochette and P.J. Ramsay, (2002), Anal. Chem., 74:5907-5911.}
}
\usage{
data(mtbe)
}
\description{
A dataset containing concentration (μg/L) and measured response (area).
}
\keyword{datasets}
