% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_function.R
\name{plotDL}
\alias{plotDL}
\title{Plots segment of calibration curve showing the estimated limits.}
\usage{
plotDL(d)
}
\arguments{
\item{d}{A tibble containing x (concentration) and y (response)}
}
\description{
Plots segment of calibration curve showing the estimated limits.
}
\examples{
data(mtbe)
plotDL(mtbe)
p <- plotDL(mtbe)
p + ## add additional labels
  ggplot2::xlab(expression(paste("Concentration (ng", " ", "g" ^ {-1}, ")")))
}
