% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_function.R
\name{plotlinDL}
\alias{plotlinDL}
\title{Plots segment of calibration curve showing the estimated linear DLs.}
\usage{
plotlinDL(d)
}
\arguments{
\item{d}{A tibble containing x (concentration) and y (response)}
}
\description{
Plots segment of calibration curve showing the estimated linear DLs.
}
\examples{
data(mtbe)
plotlinDL(mtbe)
p <- plotlinDL(mtbe)
p + ## add additional labels
  ggplot2::xlab(expression(paste("Concentration (ng", " ", "g" ^ {-1}, ")")))
}
