% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_functions.R
\name{reportDL}
\alias{reportDL}
\title{A wrapper function for reporting detection limits}
\usage{
reportDL(d, dec_point = NULL)
}
\arguments{
\item{d}{A vector containing x and y}
}
\value{

}
\description{
A wrapper function for reporting detection limits
}
\examples{
reportDL(d)
}
