% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resid_plot_function.R
\name{resid_plot}
\alias{resid_plot}
\title{Plots residual plots for different regressions}
\usage{
resid_plot(d, model_type = NULL)
}
\arguments{
\item{d}{A tibble containing x (concentration) and y (response)}

\item{model_type}{(l)inear, (q)uadratic or (p)ower regression}
}
\description{
Plots residual plots for different regressions
}
\examples{
data(chloromethane)
resid_plot(chloromethane)
resid_plot(chloromethane, "q")
}
