% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_function.R
\name{show_limits}
\alias{show_limits}
\title{Plots segment of calibration curve showing the position of the different DLs}
\usage{
show_limits(d)
}
\arguments{
\item{d}{A vector containing x and y}
}
\description{
Plots segment of calibration curve showing the position of the different DLs
}
\examples{
show_limits(d)
}
