% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_functions.R
\name{summaryDL}
\alias{summaryDL}
\title{Summarises detection limits}
\source{
{
i) J.C. Miller and J.N. Miller (1993), "Statistics for Analytical Chemistry",
3rd ed., Prentice-Hall.
ii) J. Vogelgesang and J. Hädrich (1998), Accred. Qual. Assur., 3:242-255.
iii) A. Hubaux and G. Vos (1970), Anal. Chem., 42:849-855
& D.T. O'Neill, E.A. Rochette and P.J. Ramsay, (2002), Anal. Chem., 74:5907-5911
}
}
\usage{
summaryDL(d, dp = NULL)
}
\arguments{
\item{d}{A tibble containing x (concentration) and y (response).}

\item{dp}{Number of decimal points}
}
\description{
A generic function summarising the detection limits (DLs) estimated using three approaches;
i) Miller and Miller, ii) Vogelgesang and Hädrich, and iii) Hubert & Vos.
The first two DLs are directly estimated using internal functions, \emph{dl_miller} and \emph{dl_vogelhad}
while the third DL is estimated iteratively using \emph{dl_hubertvos}.
By default, a single decimal point is shown but this can be changed by the user.
}
\examples{
data(mtbe)
summaryDL(mtbe) 	    #single decimal point
summaryDL(mtbe, 3)	#three decimal points
}
