% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulate_function.R
\name{tabulateDL}
\alias{tabulateDL}
\title{Tabulate detection limits}
\usage{
tabulateDL(d, dp = NULL)
}
\arguments{
\item{d}{A tibble containing x (concentration) and y (response).}

\item{dp}{Number of decimal points}
}
\description{
A function which tabulates the detection limits estimated using four approaches;
i) Miller and Miller, ii) Vogelgesang and Hädrich, iii) Hubert & Vos, and iv) the R \emph{chemCal} package.
By default, a single decimal point is shown but this can be changed by the user.
}
\examples{
data(mtbe)
tabulateDL(mtbe) 	 #single decimal point
tabulateDL(mtbe, 3)	#three decimal points
}
