% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_functions.R
\name{reportDL}
\alias{reportDL}
\title{Reports detection limits ()}
\usage{
reportDL(dat)
}
\arguments{
\item{dat}{A vector containing x and y}
}
\description{
Reports detection limits ()
}
\examples{
reportDL(dat)
}
